requirements.txt
flask
flask_cors
python-dotenv
google-generativeai
langchain-google-genai
langchain-community
langchain-core
langchain
chromadb